A4A signup: carry acquisition query params to conversion events
