+++ client/a8c-for-agencies/sections/signup/signup-v2/components/multi-step-form/hooks/use-submit-signup.ts
@@ -1,5 +1,6 @@
 import { useCallback } from 'react';
 import useCreateAgencyMutation from 'calypso/a8c-for-agencies/sections/signup/agency-details-form/hooks/use-create-agency-mutation';
+import useAcquisitionProps from 'calypso/a8c-for-agencies/sections/signup/hooks/use-acquisition-props';
 import { saveSignupDataToLocalStorage } from 'calypso/a8c-for-agencies/sections/signup/lib/signup-data-to-local-storage';
 import { useHandleWPCOMRedirect } from 'calypso/a8c-for-agencies/sections/signup/signup-form/hooks/use-handle-wpcom-redirect';
 import { AgencyDetailsSignupPayload } from 'calypso/a8c-for-agencies/sections/signup/types';
@@ -17,6 +18,7 @@ export default function useSubmitSignup() {
 
 	const queryParams = new URLSearchParams( window.location.search );
 	const referer = queryParams.get( 'ref' );
+	const acquisitionProps = useAcquisitionProps();
 	const userLoggedIn = useSelector( isUserLoggedIn );
 	const shouldRedirectToWPCOM = ! userLoggedIn;
 	const handleWPCOMRedirect = useHandleWPCOMRedirect();
@@ -68,9 +70,17 @@ export default function useSubmitSignup() {
 					state: payload.state,
 					referer,
 					phone_number: payload.phoneNumber ?? '',
+					...acquisitionProps,
 				} )
 			);
 		},
-		[ dispatch, shouldRedirectToWPCOM, createAgency, referer, handleWPCOMRedirect ]
+		[
+			dispatch,
+			shouldRedirectToWPCOM,
+			createAgency,
+			referer,
+			handleWPCOMRedirect,
+			acquisitionProps,
+		]
 	);
 }
