+++ client/a8c-for-agencies/sections/signup/signup-v2/components/multi-step-form/index.tsx
@@ -8,10 +8,10 @@ import A4ALogo, {
 } from 'calypso/a8c-for-agencies/components/a4a-logo';
 import { useIsDarkMode } from 'calypso/a8c-for-agencies/hooks/use-is-dark-mode';
 import { AgencyDetailsSignupPayload } from 'calypso/a8c-for-agencies/sections/signup/types';
-import { useDispatch, useSelector } from 'calypso/state';
+import { useDispatch } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { errorNotice } from 'calypso/state/notices/actions';
-import getCurrentQueryArguments from 'calypso/state/selectors/get-current-query-arguments';
+import useAcquisitionProps from '../../../hooks/use-acquisition-props';
 import useCreateSignupMutation from '../../../hooks/use-create-signup-mutation';
 import StepProgress from '../step-progress';
 import BlueprintForm from './blueprint-form';
@@ -39,15 +39,6 @@ type Step = {
 	isActive: boolean;
 	value: number;
 };
-
-const ACQUISITION_QUERY_PARAMS = [
-	'utm_source',
-	'utm_medium',
-	'utm_campaign',
-	'utm_term',
-	'utm_content',
-	'ref',
-];
 
 const STEP_NOT_STARTED = 0;
 const STEP_HALFWAY = 50;
@@ -96,20 +87,7 @@ const MultiStepForm = ( {
 	const [ currentStep, setCurrentStep ] = useState( 1 );
 	const dispatch = useDispatch();
 	const isDarkMode = useIsDarkMode();
-	const queryArgs = useSelector( getCurrentQueryArguments );
-
-	// Keys are omitted entirely when unset so we never record empty values.
-	const acquisitionProps = useMemo(
-		() =>
-			ACQUISITION_QUERY_PARAMS.reduce< Record< string, string > >( ( props, key ) => {
-				const value = queryArgs?.[ key ];
-				if ( typeof value === 'string' && value !== '' ) {
-					props[ key ] = value;
-				}
-				return props;
-			}, {} ),
-		[ queryArgs ]
-	);
+	const acquisitionProps = useAcquisitionProps();
 
 	const [ formData, setFormData ] = useState< Partial< AgencyDetailsSignupPayload > >( {} );
 
@@ -137,7 +115,12 @@ const MultiStepForm = ( {
 
 	const { mutate: submitSurvey, isPending: isSubmittingSurveyPending } = useCreateSignupMutation( {
 		onSuccess: () => {
-			dispatch( recordTracksEvent( 'calypso_a4a_agency_signup_form_via_magic_link_submitted' ) );
+			dispatch(
+				recordTracksEvent(
+					'calypso_a4a_agency_signup_form_via_magic_link_submitted',
+					acquisitionProps
+				)
+			);
 		},
 		onError: ( error: APIError ) => {
 			dispatch( errorNotice( error?.message, { id: notificationId } ) );
